Refactored the project structure
  Added unit test framework
  Also, modified the models to remove curses calls from within. Broke
  out curses rendering into a separate object. Added a terrain_map
  object (not finished yet)

  Other stuff, too
